# don't execute long operations more than once per iteration