--- a/src/BenchmarkDotNet/Engines/EngineParameters.cs
+++ b/src/BenchmarkDotNet/Engines/EngineParameters.cs
@@ -20,7 +20,6 @@ namespace BenchmarkDotNet.Engines
         public Action GlobalCleanupAction { get; set; } = null;
         public Action IterationSetupAction { get; set; } = null;
         public Action IterationCleanupAction { get; set; } = null;
-        public IResolver Resolver { get; set; }
         public bool MeasureGcStats { get; set; }
     }
 }