--- a/src/BenchmarkDotNet/Engines/IEngineFactory.cs
+++ b/src/BenchmarkDotNet/Engines/IEngineFactory.cs
@@ -2,6 +2,6 @@ namespace BenchmarkDotNet.Engines
 {
     public interface IEngineFactory
     {
-        IEngine Create(EngineParameters engineParameters);
+        IEngine CreateReadyToRun(EngineParameters engineParameters);
     }
 }