Background: because of the default unroll factor = 16, every iteration was invoking benchmark at least 16 times. 

So if the benchmark took 1s, then:

1. jitting took 16x1s
2. warmup took 6 (MinIterationCount for EngineWarmupStage)  x 16 x 1s = 96 s
3. every target iteration was taking 16s

**Note:** In the default mode, this was described in our docs and was possible to change with custom config. We just did not have the time to fix it.


This PR introduces following changes:

1. EngineFactory is supposed to create an engine which is "ready to run", which means it's supposed to call `GlobalSetup`. (hence the method name change `Create` => `CreateReadyToRun`)
2. Engine implements `IDisposable`, which calls `GlobalCleanup`
3. When the strategy requires us to jit the code first (default mode) and to run the pilot (no explicit value for invocationCount), the engine factory does jitting and **also checks how long does it take to run the benchmark once** (unrol factor=invocation count=1).
 a. If it took longer than IterationTime (0.5s by default), the engine is created for job with one invocation per iteration
 b. otherwise, old rules applies
4. Strategies which don't require jitting don't follow this path (by design)
5. I had to duplicate both the C# and IL we generate to have possibility to choose the mode and keep the Engine immutable. It does not look as bad as  I expected it to look ;)
6. Because the changes applies mostly to `EngineFactory`, I was able to write some nice unit tests (not integration tests)

```cs
public class Sleeep
{
    [Benchmark] public void Sleep1s() => Thread.Sleep(TimeSpan.FromSeconds(1));
}
```

Before: Total time: 00:06:16 (376.01 sec)
After:   Total time: 00:00:27 (27.88 sec)

This fixes #736, makes #730 easy and contributes to #550.

@AndreyAkinshin please take a look ;)
@krzysztofcwalina this fixes the issue you have faced in ML.NET