@AndreyAkinshin 2018-05-22
Currently, I have the following output for a 50ms method:
```
IdleJit -1: 1 op, 109000.00 ns, 109.0000 us/op
Jit -1: 1 op, 53255800.00 ns, 53.2558 ms/op
IdleJit -1: 16 op, 212200.00 ns, 13.2625 us/op
Jit -1: 16 op, 848930200.00 ns, 53.0581 ms/op
Pilot  1: 16 op, 854972000.00 ns, 53.4358 ms/op

IdleWarmup  1: 16 op, 600.00 ns, 37.5000 ns/op
IdleWarmup  2: 16 op, 500.00 ns, 31.2500 ns/op
IdleWarmup  3: 16 op, 600.00 ns, 37.5000 ns/op
IdleWarmup  4: 16 op, 600.00 ns, 37.5000 ns/op
IdleWarmup  5: 16 op, 500.00 ns, 31.2500 ns/op
IdleWarmup  6: 16 op, 500.00 ns, 31.2500 ns/op
```
And the following output for a 5 sec method:
```
IdleJit -1: 1 op, 363300.00 ns, 363.3000 us/op
Jit -1: 1 op, 5004330100.00 ns, 5.0043 s/op
IdleWarmup  1: 1 op, 200.00 ns, 200.0000 ns/op
IdleWarmup  2: 1 op, 200.00 ns, 200.0000 ns/op
IdleWarmup  3: 1 op, 200.00 ns, 200.0000 ns/op
IdleWarmup  4: 1 op, 300.00 ns, 300.0000 ns/op
IdleWarmup  5: 1 op, 300.00 ns, 300.0000 ns/op
IdleWarmup  6: 1 op, 300.00 ns, 300.0000 ns/op

IdleTarget  1: 1 op, 200.00 ns, 200.0000 ns/op
IdleTarget  2: 1 op, 300.00 ns, 300.0000 ns/op
IdleTarget  3: 1 op, 300.00 ns, 300.0000 ns/op
IdleTarget  4: 1 op, 200.00 ns, 200.0000 ns/op
```
It works which is great! Awesome PR! However, I suggest to improve the output:
* `IterationIndex = -1` confuses me. Could we start it with `1`?
* Maybe it makes sense to rename `IdleJit` and `Jit` to `IdleJitting` and `Jitting`?
* It would be great to align all output lines and form "pseudocolumns" for iteration name and index. In this case, the output will be much readable.