@AndreyAkinshin 2018-05-22
I also tried to run a 50ms method with the following config:
```cs
private class Config : ManualConfig
{
    public Config()
    {
        Add(Job.Default.WithUnrollFactor(4));
    }
}
```
The output:
```
IdleJit -1: 1 op, 114500.00 ns, 114.5000 us/op
Jit -1: 1 op, 53745100.00 ns, 53.7451 ms/op
IdleJit -1: 4 op, 131000.00 ns, 32.7500 us/op
Jit -1: 4 op, 216355400.00 ns, 54.0889 ms/op
Pilot  1: 4 op, 212807600.00 ns, 53.2019 ms/op
Pilot  2: 8 op, 428189900.00 ns, 53.5237 ms/op
Pilot  3: 16 op, 861407400.00 ns, 53.8380 ms/op

IdleWarmup  1: 16 op, 400.00 ns, 25.0000 ns/op
IdleWarmup  2: 16 op, 500.00 ns, 31.2500 ns/op
IdleWarmup  3: 16 op, 400.00 ns, 25.0000 ns/op
```
If we specifically ask for `UnrollFactor=4`, this value should be used. The new auto strategy for `UnrollFactor` should be applied only for unspecified value. Also it would be nice to have a test for such case.