@adamsitnik 2018-05-22
@AndreyAkinshin very good points! I totally missed the case with `unrollFactor` set in explicit way.

I have pushed a commit with fixes.

With the aligning it definitely looks better:

```log
IdleJitting  1: 1 op, 361510.84 ns, 361.5108 us/op
MainJitting  1: 1 op, 1000483249.90 ns, 1.0005 s/op
```

Was:

```log
IdleJitting -1: 1 op, 374340.49 ns, 374.3405 us/op
Jitting -1: 1 op, 1000731574.78 ns, 1.0007 s/op
```