@AndreyAkinshin 2018-05-23
```
IdleJitting  1: 1 op, 234600.00 ns, 234.6000 us/op
MainJitting  1: 1 op, 51181500.00 ns, 51.1815 ms/op

IdleJitting  1: 16 op, 240400.00 ns, 15.0250 us/op
MainJitting  1: 16 op, 812670500.00 ns, 50.7919 ms/op

Pilot  1: 16 op, 813271600.00 ns, 50.8295 ms/op

IdleWarmup  1: 16 op, 1600.00 ns, 100.0000 ns/op
IdleWarmup  2: 16 op, 400.00 ns, 25.0000 ns/op
IdleWarmup  3: 16 op, 400.00 ns, 25.0000 ns/op
IdleWarmup  4: 16 op, 200.00 ns, 12.5000 ns/op
IdleWarmup  5: 16 op, 300.00 ns, 18.7500 ns/op
IdleWarmup  6: 16 op, 300.00 ns, 18.7500 ns/op
```
Now only the `Pilot` stage is unaligned.
@adamsitnik, could you align it as well?