@AndreyAkinshin 2018-05-23
Also `MainJitting` and `IdleWarmup` have the different amount of symbols, so it would be nice to align them as well.